successive calls to PopulateModel(delete_model=False) are now noops unless delete_model=True
git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CPAnalyst@9119 1fc53939-2000-0410-845c-e8453a809027

+++ src/DataModel.py
@@ -31,8 +31,13 @@ class DataModel(Singleton):
         return str(self.obCount)+" objects in "+ \
                str(len(self.data))+" images"
                
-    def PopulateModel(self):
-        self.DeleteModel()
+    def PopulateModel(self, delete_model=False):
+        if delete_model:
+            self.DeleteModel()
+        elif not self.IsEmpty():
+            # No op if already populated
+            return
+        
         if db is None:
             logging.error("Error: No database connection!")
             return
